Two games with their own windows.
Multiple games?!?!? - 
Utilizing a table to store each buffered game, and a variable that points to each game in the table, we can create a simple update function that only updates the selected game. 

Selecting windows -
This is fairly easy, simply checking if the user has clicked with the bounds of another window.
The issue with this however is that if you click at the same location as the window behind it, it will swap the selections
meaning you cant build blocks if your window is on top of another.
To fix this is simple but may not be obvious at first, just check if the mouse is ONLY in 1 screen and have the if statements be in chronological order, this can be achieved by having 2 statements have one check 2, then 1. then the other check 1, then 2.

Moving windows - 
Using a mouse drag event on the window that is selected we can change the windows render offsets to the point where the mouse is dragging them too, and redrawing the other window as we get closer to the point (a more advanced version of this would be to only re-draw the other window when the selected window was on top of it then moved away)